incorporate team comments
